+++ crd/multitenancy/embed_test.go
@@ -32,3 +32,29 @@ func TestGetNodeInfo(t *testing.T) {
 	_, err := GetNodeInfo()
 	assert.NoError(t, err)
 }
+
+const podNetworkFilename = "manifests/public.acn.azure.com_podnetworks.yaml"
+
+func TestEmbedPodNetwork(t *testing.T) {
+	b, err := os.ReadFile(podNetworkFilename)
+	assert.NoError(t, err)
+	assert.Equal(t, b, PodNetworkYAML)
+}
+
+func TestGetPodNetworks(t *testing.T) {
+	_, err := GetPodNetworks()
+	assert.NoError(t, err)
+}
+
+const podNetworkInstanceFilename = "manifests/public.acn.azure.com_podnetworkinstances.yaml"
+
+func TestEmbedPodNetworkInstance(t *testing.T) {
+	b, err := os.ReadFile(podNetworkInstanceFilename)
+	assert.NoError(t, err)
+	assert.Equal(t, b, PodNetworkInstanceYAML)
+}
+
+func TestGetPodNetworkInstances(t *testing.T) {
+	_, err := GetPodNetworkInstances()
+	assert.NoError(t, err)
+}
