--- crd/multitenancy/external/manifests/doc.go
@@ -1,3 +1,0 @@
-// Package manifests exists to allow the rendered CRD manifests to be
-// packaged in to dependent components.
-package manifests
