+++ crd/multitenancy/manifests/public.acn.azure.com_podnetworks.yaml
@@ -3,7 +3,8 @@ apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
 metadata:
   annotations:
-    controller-gen.kubebuilder.io/version: v0.12.0
+    controller-gen.kubebuilder.io/version: v0.11.3
+  creationTimestamp: null
   name: podnetworks.public.acn.azure.com
 spec:
   group: public.acn.azure.com
